check for non-zero size file, not just file
Curl creates a zero size file when a download starts, so checking for the existence of a file isn't sufficient for making sure a download occurred. When curl timesout, it leaves behind a zero-size file that was messing up checks. Checking for a file with non-zero size is the test we wanted.